+++ routerd/src/zebra/linux/rtnetlink.rs
@@ -21,6 +21,10 @@
 */
 
 use std::mem::size_of;
+use std::ptr::copy;
+
+use super::netlink::*;
+
 
 const RTA_ALIGNTO: usize = 4usize;
 
@@ -41,5 +45,71 @@ pub fn rta_data() -> usize {
 pub struct Rtattr {
     pub rta_len: u16,
     pub rta_type: u16,
+}
+
+
+fn addattr_ptr(ptr: *const libc::c_void, offset: usize,
+               rta_type: libc::c_int, rta_len: usize, src_ptr: *const libc::c_void, alen: usize) {
+    unsafe {
+        let rta = ptr.offset(offset as isize) as *mut Rtattr;
+        let rta_ptr = rta as *const _ as *mut libc::c_void;
+        let dst_ptr = rta_ptr.offset(size_of::<Rtattr>() as isize);
+
+        (*rta).rta_len = rta_len as u16;
+        (*rta).rta_type = rta_type as u16;
+
+        copy(src_ptr, dst_ptr, alen);
+    }
+}
+
+pub fn addattr_l(h: &mut Nlmsghdr, maxlen: usize, rta_type: libc::c_int, src: &[u8], alen: usize) -> bool {
+    let rta_len = rta_length(alen);
+    let offset = nlmsg_align(h.nlmsg_len as usize);
+
+    if offset + rta_len > maxlen {
+        false
+    } else {
+        let ptr = h as *const _ as *const libc::c_void;
+        let src_ptr = src as *const _ as *const libc::c_void;
+
+        addattr_ptr(ptr, offset, rta_type, rta_len, src_ptr, alen);
+        h.nlmsg_len = (offset + rta_len) as u32;
+
+        true
+    }
+}
+
+pub fn rta_addattr_l(rta: &mut Rtattr, maxlen: usize, rta_type: libc::c_int, src: &[u8], alen: usize) -> bool {
+    let rta_len = rta_length(alen);
+    let offset = rta_align(rta.rta_len as usize);
+
+    if offset + rta_len > maxlen {
+        false
+    } else {
+        let ptr = rta as *const _ as *const libc::c_void;
+        let src_ptr = src as *const _ as *const libc::c_void;
+
+        addattr_ptr(ptr, offset, rta_type, rta_len, src_ptr, alen);
+        rta.rta_len = (offset + rta_len) as u16;
+
+        true
+    }
+}
+
+pub fn addattr32(h: &mut Nlmsghdr, maxlen: usize, rta_type: libc::c_int, src: u32) -> bool {
+    let rta_len = rta_length(size_of::<u32>());
+    let offset = nlmsg_align(h.nlmsg_len as usize);
+
+    if offset + rta_len > maxlen {
+        false
+    } else {
+        let ptr = h as *const _ as *const libc::c_void;
+        let src_ptr = &src as *const _ as *mut libc::c_void;
+
+        addattr_ptr(ptr, offset, rta_type, rta_len, src_ptr, size_of::<u32>());
+        h.nlmsg_len = (offset + rta_len) as u32;
+
+        true
+    }
 }
